Year 2023, Day 18
For part 1, I started off tracking the actual grid, and used a
quick'n'dirty recursive flood fill from just outside the "known world"
to determine what the inner area is. This was with the expectation
that something was going to happen with the colors in part 2.

Cue the AOC curve ball!

For part 2, I threw out almost everything except the input parsing
(and had to adjust that as well, of course) and then used a slightly
tweaked version of the shoelace formula, adapted in a way to account
for the fact that these are lines with actual "width" and not
imaginary 0 width lines.