Part 2 of Year 2023, Day 20
The solution written here is 100% geared toward the specific input
file I received, which had a single conjunction module sitting in
front of the 'rx' target (noted after visualizing with graph.py).

After eyeballing the graph and instrumenting the code to specifically
look for when that conjunction module was receiving high pulses, it
was clear there was a repeating pattern of varying lenghts for each of
them.

The next step was to track specifically those inputs, and aborting
once each of them had seen at least one repeat, and then multiplying
the resulting values.